fix(conversation): sync chat-list + in-chat status dots with live presence
The conversation-mode status dot in the chat list (ChatSidebar) and the
in-chat character overlay (ChatArea) read the per-chat
`conversationCharacterStatuses` snapshot, which is only rewritten during
message generation. Setting a manual presence override (or a schedule
transition) updated the toolbar presence pill — which derives status live
from override > schedule > default — but left the sidebar dot showing the
stale generation-time status until the next generated message.

Extract the status-derivation helpers (getCurrentStatus,
getActiveStatusOverride, getEffectiveCurrentStatus) and their schedule types
into @marinara-engine/shared so the client derives presence the same way the
server does, and have the sidebar + in-chat overlay prefer the live
override/schedule status over the snapshot (falling back to the snapshot when
there is no active override or schedule). schedule.service.ts re-exports the
moved helpers so existing server imports are unchanged.
