+++ packages/server/src/services/conversation/schedule.service.ts
@@ -6,53 +6,30 @@
 
 import { createLLMProvider } from "../llm/provider-registry.js";
 import type { BaseLLMProvider } from "../llm/base-provider.js";
-import type { ConversationPresenceStatus, ConversationStatusOverride } from "@marinara-engine/shared";
+import {
+  CONVERSATION_SCHEDULE_DAYS,
+  getActiveStatusOverride,
+  getCurrentStatus,
+  getEffectiveCurrentStatus,
+  type CharacterSchedules,
+  type ConversationPresenceStatus,
+  type ConversationStatusOverride,
+  type CurrentConversationStatus,
+  type DaySchedule,
+  type ScheduleBlock,
+  type WeekSchedule,
+} from "@marinara-engine/shared";
 
-// ── Types ──
-
-/** A single time block in a character's daily schedule */
-export interface ScheduleBlock {
-  /** Hour range, e.g. "06:00-08:00" */
-  time: string;
-  /** What the character is doing */
-  activity: string;
-  /** Derived status for this block */
-  status: ConversationPresenceStatus;
-}
-
-/** One day of a character's schedule */
-export type DaySchedule = ScheduleBlock[];
-
-/** Full weekly schedule for a character */
-export interface WeekSchedule {
-  /** ISO date string of the Monday this schedule starts */
-  weekStart: string;
-  /** Schedules keyed by day name */
-  days: Record<string, DaySchedule>;
-  /** How many minutes of user inactivity before this character messages unprompted (0 = never) */
-  inactivityThresholdMinutes: number;
-  /** Optional exact response delay in minutes while idle */
-  idleResponseDelayMinutes?: number;
-  /** Optional exact response delay in minutes while busy / DND */
-  dndResponseDelayMinutes?: number;
-  /** How chatty the character is — affects autonomous messaging frequency (0-100) */
-  talkativeness: number;
-}
-
-/** All character schedules stored in chat metadata */
-export interface CharacterSchedules {
-  [characterId: string]: WeekSchedule;
-}
-
-export interface CurrentConversationStatus {
-  status: ConversationPresenceStatus;
-  activity: string;
-  override?: ConversationStatusOverride;
-}
+// The schedule/override status-derivation helpers and their schedule types now
+// live in @marinara-engine/shared so the client presence dots derive status the
+// same way the server does. Re-export them here so existing server imports from
+// "./schedule.service.js" keep resolving unchanged.
+export { getActiveStatusOverride, getCurrentStatus, getEffectiveCurrentStatus };
+export type { CharacterSchedules, CurrentConversationStatus, DaySchedule, ScheduleBlock, WeekSchedule };
 
 // ── Constants ──
 
-const DAYS = ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"];
+const DAYS = CONVERSATION_SCHEDULE_DAYS;
 
 const STATUS_KEYWORDS: Record<string, ConversationPresenceStatus> = {
   sleep: "offline",
@@ -275,72 +252,8 @@ function inferStatusFromActivity(activity: string): ConversationPresenceStatus {
 }
 
 // ── Status Derivation ──
-
-/**
- * Get the current status and activity for a character based on their schedule.
- */
-export function getCurrentStatus(
-  schedule: WeekSchedule,
-  now: Date = new Date(),
-): { status: ConversationPresenceStatus; activity: string } {
-  const dayName = DAYS[(now.getDay() + 6) % 7]!; // JS Sunday=0, we want Monday=0
-  const daySchedule = schedule.days[dayName];
-  if (!daySchedule || daySchedule.length === 0) {
-    return { status: "online", activity: "free time" };
-  }
-
-  const currentMinutes = now.getHours() * 60 + now.getMinutes();
-
-  for (const block of daySchedule) {
-    const [startStr, endStr] = block.time.split("-");
-    if (!startStr || !endStr) continue;
-
-    const [sh, sm] = startStr.split(":").map(Number);
-    const [eh, em] = endStr.split(":").map(Number);
-    const startMin = (sh ?? 0) * 60 + (sm ?? 0);
-    const endMin = (eh ?? 0) * 60 + (em ?? 0);
-
-    // Handle blocks that don't wrap around midnight
-    if (startMin <= currentMinutes && currentMinutes < endMin) {
-      return { status: block.status, activity: block.activity };
-    }
-    // Handle midnight-wrapping blocks (e.g., 23:00-07:00)
-    if (startMin > endMin && (currentMinutes >= startMin || currentMinutes < endMin)) {
-      return { status: block.status, activity: block.activity };
-    }
-  }
-
-  return { status: "online", activity: "free time" };
-}
-
-function isManualPresenceStatus(value: unknown): value is ConversationStatusOverride["status"] {
-  return value === "online" || value === "idle" || value === "dnd" || value === "offline";
-}
-
-export function getActiveStatusOverride(
-  override: ConversationStatusOverride | null | undefined,
-  now: Date = new Date(),
-): ConversationStatusOverride | null {
-  if (!override || !isManualPresenceStatus(override.status)) return null;
-  if (typeof override.expiresAt === "string") {
-    const expiresAt = new Date(override.expiresAt).getTime();
-    if (!override.expiresAt.trim() || !Number.isFinite(expiresAt) || expiresAt <= now.getTime()) return null;
-  }
-  return override;
-}
-
-export function getEffectiveCurrentStatus(
-  schedule: WeekSchedule | null | undefined,
-  override: ConversationStatusOverride | null | undefined,
-  now: Date = new Date(),
-  fallbackActivity = "free time",
-): CurrentConversationStatus {
-  const scheduled = schedule ? getCurrentStatus(schedule, now) : { status: "online" as const, activity: fallbackActivity };
-  const activeOverride = getActiveStatusOverride(override, now);
-  if (!activeOverride) return scheduled;
-  const activity = typeof activeOverride.activity === "string" ? activeOverride.activity.trim() : scheduled.activity;
-  return { status: activeOverride.status, activity, override: activeOverride };
-}
+// getCurrentStatus / getActiveStatusOverride / getEffectiveCurrentStatus moved to
+// @marinara-engine/shared (utils/conversation-presence) — imported + re-exported above.
 
 /**
  * Get a human-readable summary of today's schedule for a character.
