+++ packages/shared/src/index.ts
@@ -95,3 +95,4 @@ export * from "./utils/thinking-tags.js";
 export * from "./utils/lorebook-folder-tree.js";
 export * from "./utils/text-matching.js";
 export * from "./utils/sprite-labels.js";
+export * from "./utils/conversation-presence.js";
